pyproject.toml
[project]
name = "django-jinja-render-block"
version = "0.3.0"
description = "Render the content of a specific block tag from a Jinja2 template."
authors = [
    {name = "Kevin Renskers", email = "kevin@loopwerk.io"},
]
license = "MIT"
license-files = [ "LICENSE" ]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django-jinja>=2.11.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
]

[project.urls]
Homepage = "https://github.com/loopwerk/django-jinja-render-block/"
Repository = "https://github.com/loopwerk/django-jinja-render-block.git"
Issues = "https://github.com/loopwerk/django-jinja-render-block/issues"

[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "render_block"
module-root = ""

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["render_block"]

[tool.ruff]
line-length = 120
lint.extend-select = ["I", "N"]

[tool.mypy]
disable_error_code = ["import-untyped"]
warn_redundant_casts = true
check_untyped_defs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = "test*.py"
filterwarnings = ["ignore::DeprecationWarning"]